# src/tsoa/swagger.json
{
	"components": {
...
		"schemas": {
			"AuthResponse": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FailureResponse": {
				"properties": {
					"error": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Buildings": {
				"properties": {
					"planetID": {
						"type": "number",
						"format": "double"
					},
					"metalMine": {
						"type": "number",
						"format": "double"
					},
					"crystalMine": {
						"type": "number",
						"format": "double"
					},
					"deuteriumSynthesizer": {
						"type": "number",
						"format": "double"
					},
					"solarPlant": {
						"type": "number",
						"format": "double"
					},
					"fusionReactor": {
						"type": "number",
						"format": "double"
					},
					"roboticFactory": {
						"type": "number",
						"format": "double"
					},
					"naniteFactory": {
						"type": "number",
						"format": "double"
					},
					"shipyard": {
						"type": "number",
						"format": "double"
					},
					"metalStorage": {
						"type": "number",
						"format": "double"
					},
					"crystalStorage": {
						"type": "number",
						"format": "double"
					},
					"deuteriumStorage": {
						"type": "number",
						"format": "double"
					},
					"researchLab": {
						"type": "number",
						"format": "double"
					},
					"terraformer": {
						"type": "number",
						"format": "double"
					},
					"allianceDepot": {
						"type": "number",
						"format": "double"
					},
					"missileSilo": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"planetID",
					"metalMine",
					"crystalMine",
					"deuteriumSynthesizer",
					"solarPlant",
					"fusionReactor",
					"roboticFactory",
					"naniteFactory",
					"shipyard",
					"metalStorage",
					"crystalStorage",
					"deuteriumStorage",
					"researchLab",
					"terraformer",
					"allianceDepot",
					"missileSilo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Globals.PlanetType": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Planet": {
				"properties": {
					"planetID": {
						"type": "number",
						"format": "double"
					},
					"ownerID": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"posGalaxy": {
						"type": "number",
						"format": "double"
					},
					"posSystem": {
						"type": "number",
						"format": "double"
					},
					"posPlanet": {
						"type": "number",
						"format": "double"
					},
					"lastUpdate": {
						"type": "number",
						"format": "double"
					},
					"planetType": {
						"$ref": "#/components/schemas/Globals.PlanetType"
					},
					"image": {
						"type": "string"
					},
					"diameter": {
						"type": "number",
						"format": "double"
					},
					"fieldsCurrent": {
						"type": "number",
						"format": "double"
					},
					"fieldsMax": {
						"type": "number",
						"format": "double"
					},
					"tempMin": {
						"type": "number",
						"format": "double"
					},
					"tempMax": {
						"type": "number",
						"format": "double"
					},
					"metal": {
						"type": "number",
						"format": "double"
					},
					"crystal": {
						"type": "number",
						"format": "double"
					},
					"deuterium": {
						"type": "number",
						"format": "double"
					},
					"energyUsed": {
						"type": "number",
						"format": "double"
					},
					"energyMax": {
						"type": "number",
						"format": "double"
					},
					"metalMinePercent": {
						"type": "number",
						"format": "double"
					},
					"crystalMinePercent": {
						"type": "number",
						"format": "double"
					},
					"deuteriumSynthesizerPercent": {
						"type": "number",
						"format": "double"
					},
					"solarPlantPercent": {
						"type": "number",
						"format": "double"
					},
					"fusionReactorPercent": {
						"type": "number",
						"format": "double"
					},
					"solarSatellitePercent": {
						"type": "number",
						"format": "double"
					},
					"bBuildingId": {
						"type": "number",
						"format": "double"
					},
					"bBuildingEndTime": {
						"type": "number",
						"format": "double"
					},
					"bBuildingDemolition": {
						"type": "boolean"
					},
					"bHangarQueue": {
						"type": "string"
					},
					"bHangarStartTime": {
						"type": "number",
						"format": "double"
					},
					"bHangarPlus": {
						"type": "boolean"
					},
					"destroyed": {
						"type": "boolean"
					}
				},
				"required": [
					"planetID",
					"ownerID",
					"name",
					"posGalaxy",
					"posSystem",
					"posPlanet",
					"lastUpdate",
					"planetType",
					"image",
					"diameter",
					"fieldsCurrent",
					"fieldsMax",
					"tempMin",
					"tempMax",
					"metal",
					"crystal",
					"deuterium",
					"energyUsed",
					"energyMax",
					"metalMinePercent",
					"crystalMinePercent",
					"deuteriumSynthesizerPercent",
					"solarPlantPercent",
					"fusionReactorPercent",
					"solarSatellitePercent",
					"bBuildingId",
					"bBuildingEndTime",
					"bBuildingDemolition",
					"bHangarQueue",
					"bHangarStartTime",
					"bHangarPlus",
					"destroyed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BuildBuildingRequest": {
				"properties": {
					"planetID": {
						"type": "number",
						"format": "double"
					},
					"buildingID": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"planetID",
					"buildingID"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CancelBuildingRequest": {
				"properties": {
					"planetID": {
						"type": "number",
						"format": "double"
					},
					"buildingID": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"planetID",
					"buildingID"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DemolishBuildingRequest": {
				"properties": {
					"planetID": {
						"type": "number",
						"format": "double"
					},
					"buildingID": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"planetID",
					"buildingID"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResourceCollection": {
				"properties": {
					"metal": {
						"type": "number",
						"format": "double"
					},
					"crystal": {
						"type": "number",
						"format": "double"
					},
					"deuterium": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"metal",
					"crystal",
					"deuterium"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStartPlanet": {
				"properties": {
					"name": {
						"type": "string"
					},
					"diameter": {
						"type": "number",
						"format": "double"
					},
					"fields": {
						"type": "number",
						"format": "double"
					},
					"resources": {
						"$ref": "#/components/schemas/IResourceCollection"
					},
					"minPlanetPos": {
						"type": "number",
						"format": "double"
					},
					"maxPlanetPos": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"diameter",
					"fields",
					"resources",
					"minPlanetPos",
					"maxPlanetPos"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ILimit": {
				"properties": {
					"min": {
						"type": "number",
						"format": "double"
					},
					"max": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"min",
					"max"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ILimits": {
				"properties": {
					"galaxy": {
						"$ref": "#/components/schemas/ILimit"
					},
					"system": {
						"$ref": "#/components/schemas/ILimit"
					},
					"position": {
						"$ref": "#/components/schemas/ILimit"
					}
				},
				"required": [
					"galaxy",
					"system",
					"position"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IServer": {
				"properties": {
					"speed": {
						"type": "number",
						"format": "double"
					},
					"language": {
						"type": "string"
					},
					"startPlanet": {
						"$ref": "#/components/schemas/IStartPlanet"
					},
					"limits": {
						"$ref": "#/components/schemas/ILimits"
					}
				},
				"required": [
					"speed",
					"language",
					"startPlanet",
					"limits"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICosts": {
				"properties": {
					"metal": {
						"type": "number",
						"format": "double"
					},
					"crystal": {
						"type": "number",
						"format": "double"
					},
					"deuterium": {
						"type": "number",
						"format": "double"
					},
					"energy": {
						"type": "number",
						"format": "double"
					},
					"factor": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"metal",
					"crystal",
					"deuterium",
					"energy",
					"factor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRequirement": {
				"properties": {
					"unitID": {
						"type": "number",
						"format": "double"
					},
					"level": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"unitID",
					"level"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBuilding": {
				"properties": {
					"unitID": {
						"type": "number",
						"format": "double"
					},
					"costs": {
						"$ref": "#/components/schemas/ICosts"
					},
					"requirements": {
						"items": {
							"$ref": "#/components/schemas/IRequirement"
						},
						"type": "array"
					}
				},
				"required": [
					"unitID",
					"costs",
					"requirements"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITechnology": {
				"properties": {
					"unitID": {
						"type": "number",
						"format": "double"
					},
					"costs": {
						"$ref": "#/components/schemas/ICosts"
					},
					"requirements": {
						"items": {
							"$ref": "#/components/schemas/IRequirement"
						},
						"type": "array"
					}
				},
				"required": [
					"unitID",
					"costs",
					"requirements"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStats": {
				"properties": {
					"consumption": {
						"type": "number",
						"format": "double"
					},
					"speed": {
						"type": "number",
						"format": "double"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"consumption",
					"speed",
					"capacity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRapidfire": {
				"properties": {
					"unitID": {
						"type": "number",
						"format": "double"
					},
					"amount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"unitID",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IShip": {
				"properties": {
					"unitID": {
						"type": "number",
						"format": "double"
					},
					"costs": {
						"$ref": "#/components/schemas/ICosts"
					},
					"requirements": {
						"items": {
							"$ref": "#/components/schemas/IRequirement"
						},
						"type": "array"
					},
					"stats": {
						"$ref": "#/components/schemas/IStats"
					},
					"rapidfire": {
						"items": {
							"$ref": "#/components/schemas/IRapidfire"
						},
						"type": "array"
					}
				},
				"required": [
					"unitID",
					"costs",
					"requirements",
					"stats",
					"rapidfire"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IDefense": {
				"properties": {
					"unitID": {
						"type": "number",
						"format": "double"
					},
					"costs": {
						"$ref": "#/components/schemas/ICosts"
					},
					"requirements": {
						"items": {
							"$ref": "#/components/schemas/IRequirement"
						},
						"type": "array"
					},
					"stats": {
						"$ref": "#/components/schemas/IStats"
					},
					"rapidfire": {
						"items": {
							"$ref": "#/components/schemas/IRapidfire"
						},
						"type": "array"
					}
				},
				"required": [
					"unitID",
					"costs",
					"requirements",
					"stats",
					"rapidfire"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUnits": {
				"properties": {
					"buildings": {
						"items": {
							"$ref": "#/components/schemas/IBuilding"
						},
						"type": "array"
					},
					"technologies": {
						"items": {
							"$ref": "#/components/schemas/ITechnology"
						},
						"type": "array"
					},
					"ships": {
						"items": {
							"$ref": "#/components/schemas/IShip"
						},
						"type": "array"
					},
					"defenses": {
						"items": {
							"$ref": "#/components/schemas/IDefense"
						},
						"type": "array"
					}
				},
				"required": [
					"buildings",
					"technologies",
					"ships",
					"defenses"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGameConfig": {
				"properties": {
					"server": {
						"$ref": "#/components/schemas/IServer"
					},
					"units": {
						"$ref": "#/components/schemas/IUnits"
					}
				},
				"required": [
					"server",
					"units"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "access-token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "ugamela api",
		"version": "1.0.0",
		"description": "bla",
		"license": {
			"name": "AGPL-3.0"
		},
		"contact": {
			"email": "api@ugamela.org",
			"name": "the Administrator"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/login": {
			"post": {
				"operationId": "Authenticate",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthRequest"
							}
						}
					}
				}
			}
		},
		"/buildings/{planetID}": {
			"get": {
				"operationId": "GetAllBuildingsOnPlanet",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Buildings"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Buildings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "planetID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/buildings/build": {
			"post": {
				"operationId": "StartBuilding",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Planet"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Buildings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BuildBuildingRequest"
							}
						}
					}
				}
			}
		},
		"/buildings/cancel": {
			"post": {
				"operationId": "CancelBuilding",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Planet"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Buildings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CancelBuildingRequest"
							}
						}
					}
				}
			}
		},
		"/buildings/demolish": {
			"post": {
				"operationId": "DemolishBuilding",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Planet"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Buildings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DemolishBuildingRequest"
							}
						}
					}
				}
			}
		},
		"/config/game": {
			"get": {
				"operationId": "GetGameConfig",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGameConfig"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Configuration"
				],
				"security": [],
				"parameters": []
			}
		},
		"/config/units": {
			"get": {
				"operationId": "GetUnitsConfig",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUnits"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Configuration"
				],
				"security": [],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "http://127.0.0.1:3000/v1"
		}
	]
}